Refactor: merge related test modules
